feat: MainActivity로 이동
