+++ app/src/main/AndroidManifest.xml
@@ -12,20 +12,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Sunshine"
         tools:targetApi="31">
-        <!--        <activity-->
-        <!--            android:name=".MainActivity"-->
-        <!--            android:exported="true"-->
-        <!--            android:label="@string/app_name"-->
-        <!--            android:theme="@style/Theme.Sunshine">-->
-        <!--            <intent-filter>-->
-        <!--                <action android:name="android.intent.action.MAIN" />-->
-
-        <!--                <category android:name="android.intent.category.LAUNCHER" />-->
-        <!--            </intent-filter>-->
-        <!--        </activity>-->
-
         <activity
-            android:name=".presention.HomeActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Sunshine">
@@ -36,5 +24,4 @@
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
